#4 [FEAT] Double -> Float 변환 & case 분류 후 소수점 버리기 구현

+++ SwiftUI/ReviewWQuestion-Calculator/Views/fifthRow.swift
@@ -26,7 +26,7 @@ struct fifthRow: View {
                     resultString += "0"
                 }
                 
-                calculateData.result = Double(resultString) ?? 0
+                calculateData.result = Float(resultString) ?? 0
                 
             }) {
                 Text("0")
